merge part2 into master

+++ src/main/scala/ufl/RestInterface.scala
@@ -7,12 +7,18 @@ import spray.httpx.SprayJsonSupport._
 import spray.json.DefaultJsonProtocol._
 import spray.routing._
 import spray.json._
-
+import scala.util.Random
 import scala.concurrent.duration._
 import scala.language.postfixOps
 import ufl.FacebookAPI._
 import scala.collection.immutable.Map
 import java.util.Calendar
+import java.security._
+import javax.crypto._
+import javax.crypto.SecretKey
+import javax.crypto.spec.SecretKeySpec
+import java.security.spec.X509EncodedKeySpec
+import java.util.Base64
 
 class RestInterface extends HttpServiceActor
   with RestApi {
@@ -79,37 +85,56 @@ trait RestApi extends HttpService with ActorLogging { actor: Actor =>
           entity(as[FBPost]) { post => requestContext =>
             val responder = createResponder(requestContext)
             val resultUser: Option[UserNode] = RestApi.userList.find(_.id == post.userId)
-
             if(resultUser.isEmpty) {
               //invalid user id
               responder ! NodeNotFound("User")
             } else {
+              if (authenticate(post.userId, post.auth)) { 
               //Valid user, create a new post.
-              var newPostId = RestApi.getId
-              val postNode: PostNode = new PostNode(newPostId, resultUser.get.id, post.content)
-              RestApi.postList = RestApi.postList :+ postNode
-              resultUser.get.postList = resultUser.get.postList :+ postNode.id
-              println("Created new post by: " + resultUser.get.id + ", id: " + postNode.id + ", posts: " + resultUser.get.postList.length)
-              responder ! NodeCreated(newPostId)
-            }
+                var newPostId = RestApi.getId
+                val postNode: PostNode = new PostNode(newPostId, resultUser.get.id, post.content, post.authUsers)
+                RestApi.postList = RestApi.postList :+ postNode
+                resultUser.get.postList = resultUser.get.postList :+ postNode.id
+                println("[SERVER] Created new post by userId: " + resultUser.get.id + ", postId: " + postNode.id + ", total posts by user: " + resultUser.get.postList.length)
+                responder ! NodeCreated(newPostId)
+              } else {
+                responder ! AuthFailed
+              }
+            } 
           }
         }
       } ~
       path(Segment) { id =>
-        delete { requestContext =>
-          val responder = createResponder(requestContext)
-          var postDeleted = deletePost(id)
-          if(postDeleted)
-            responder ! PostDeleted
-          else
-            responder ! NodeNotFound("Post")
-        } ~
-        get { requestContext =>
-          println("get post " + id)
-          var resultPost: Option[PostNode] = RestApi.postList.find(_.id == id)
-          val responder = createResponder(requestContext)
-          resultPost.map(responder ! _.postResponse())
-            .getOrElse(responder ! NodeNotFound("Post"))
+        delete { 
+          parameter('requesterId, 'auth) { (requesterId, auth) =>
+            requestContext =>
+            val responder = createResponder(requestContext)
+            if (authenticate(requesterId, auth)) {
+              var postDeleted = deletePost(id)
+              if(postDeleted) {
+                println("[SERVER] post deleted")
+                responder ! PostDeleted
+              }
+              else
+                responder ! NodeNotFound("Post")
+            }
+            else {
+              responder ! AuthFailed
+            }
+          }
+        }~
+        get{
+          parameter('requesterId, 'auth) { (requesterId, auth) =>
+            requestContext =>
+            val responder = createResponder(requestContext)
+            if (authenticate(requesterId, auth)) {
+              var resultPost: Option[PostNode] = RestApi.postList.find(_.id == id)
+              resultPost.map(responder ! _.postResponse(requesterId))
+                .getOrElse(responder ! NodeNotFound("Post"))
+            } else {
+              responder ! AuthFailed
+            }
+          }
         }
       }
     } ~
@@ -124,10 +149,10 @@ trait RestApi extends HttpService with ActorLogging { actor: Actor =>
             } else {
               var newUserId = RestApi.getId
               val userNode: UserNode = new UserNode(newUserId, user.email, user.firstname, 
-                user.lastname, user.gender)
+                user.lastname, user.gender, user.publicKey)
               RestApi.userList = RestApi.userList :+ userNode
 
-              println("Created new user with id: " + userNode.id)
+              println("[SERVER] Created new user")
               responder ! NodeCreated(newUserId)
             }
           }
@@ -144,7 +169,6 @@ trait RestApi extends HttpService with ActorLogging { actor: Actor =>
             responder ! NodeNotFound("User")
         } ~
         get { requestContext =>
-          println("get user " + id)
           var resultUser: Option[UserNode] = RestApi.userList.find(_.id == id)
           val responder = createResponder(requestContext)
           resultUser.map(responder ! _.userResponse())
@@ -163,39 +187,60 @@ trait RestApi extends HttpService with ActorLogging { actor: Actor =>
             if (user.isEmpty) {
               responder ! NodeNotFound("User")
             } else {
-              //find if that user has the album id
-              var photoAlbumId: Option[String] = user.get.albumList.find(_ == photo.albumId)
-              if (photoAlbumId.isEmpty) {
-                responder ! NodeNotFound("Album")
-              }
-              else {
-                var newPhotoId = RestApi.getId
-                var resultAlbum: Option[AlbumNode] = RestApi.albumList.find(_.id == photoAlbumId.get)
-                val photoNode: PhotoNode = new PhotoNode(newPhotoId, photo.caption, photo.albumId, photo.creatorId, photo.photo)
-                RestApi.photoList = RestApi.photoList :+ photoNode
-                resultAlbum.get.photos = resultAlbum.get.photos :+ photoNode.id
-                responder ! NodeCreated(newPhotoId)
+              if(authenticate(photo.creatorId, photo.auth)) {
+                //user authenticated
+                //find if that user has the album id
+                var photoAlbumId: Option[String] = user.get.albumList.find(_ == photo.albumId)
+                if (photoAlbumId.isEmpty) {
+                  responder ! NodeNotFound("Album")
+                }
+                else {
+                  var newPhotoId = RestApi.getId
+                  var resultAlbum: Option[AlbumNode] = RestApi.albumList.find(_.id == photoAlbumId.get)
+                  val photoNode: PhotoNode = new PhotoNode(newPhotoId, photo.caption, photo.albumId,
+                  photo.creatorId, photo.photo, photo.authUsers)
+                  RestApi.photoList = RestApi.photoList :+ photoNode
+                  resultAlbum.get.photos = resultAlbum.get.photos :+ photoNode.id
+                  println("[SERVER] New photo uploaded with id " + newPhotoId)
+                  responder ! NodeCreated(newPhotoId)
+                }
+              } else {
+                responder ! AuthFailed
               }
             }
           }
         }
       } ~
       path(Segment) { id =>
-        delete { requestContext =>
-          val responder = createResponder(requestContext)
-          var resultPhoto: Option[PhotoNode] = RestApi.photoList.find(_.id == id)
-          if(resultPhoto.isEmpty) {
-            responder ! NodeNotFound("Photo")
-          } else {
-            RestApi.photoList = RestApi.photoList.filterNot(_.id == id)
-            responder ! PhotoDeleted
+        delete { 
+          parameter('requesterId, 'auth) { (requesterId, auth) =>
+            requestContext =>
+            val responder = createResponder(requestContext)
+            if (authenticate(requesterId, auth)) {
+              var resultPhoto: Option[PhotoNode] = RestApi.photoList.find(_.id == id)
+              if(resultPhoto.isEmpty) {
+                responder ! NodeNotFound("Photo")
+              } else {
+                RestApi.photoList = RestApi.photoList.filterNot(_.id == id)
+                responder ! PhotoDeleted
+              }
+            } else {
+              responder ! AuthFailed
+            }
           }
         } ~
-        get { requestContext =>
-          var resultPhoto: Option[PhotoNode] = RestApi.photoList.find(_.id == id)
-          val responder = createResponder(requestContext)
-          resultPhoto.map(responder ! _.photoResponse())
-            .getOrElse(responder ! NodeNotFound("Photo"))
+        get{
+          parameter('requesterId, 'auth) { (requesterId, auth) =>
+            requestContext =>
+            val responder = createResponder(requestContext)
+            if (authenticate(requesterId, auth)) {
+              var resultPhoto: Option[PhotoNode] = RestApi.photoList.find(_.id == id)
+              resultPhoto.map(responder ! _.photoResponse(requesterId))
+                .getOrElse(responder ! NodeNotFound("Photo"))
+            } else {
+              responder ! AuthFailed
+            }
+          }
         }
       }
     } ~
@@ -209,48 +254,79 @@ trait RestApi extends HttpService with ActorLogging { actor: Actor =>
             if(resultUser.isEmpty) {
               responder ! NodeNotFound("User")
             } else {
-              val albumNode: AlbumNode = new AlbumNode(newAlbumId, album.name, album.caption, 
-              album.creatorId, Calendar.getInstance().getTime().toString)
-              RestApi.albumList = RestApi.albumList :+ albumNode
-              resultUser.get.albumList = resultUser.get.albumList :+ newAlbumId
-              responder ! NodeCreated(newAlbumId)
+              if(authenticate(album.creatorId, album.auth)) {
+                val albumNode: AlbumNode = new AlbumNode(newAlbumId, album.name, album.caption, 
+                  album.creatorId, Calendar.getInstance().getTime().toString, album.authUsers)
+                RestApi.albumList = RestApi.albumList :+ albumNode
+                resultUser.get.albumList = resultUser.get.albumList :+ newAlbumId
+                println("[SERVER] new album created")
+                responder ! NodeCreated(newAlbumId)
+              } else {
+                responder ! AuthFailed
+              }
             }
           }
         }
       } ~
       path(Segment) {  id =>
-        delete { requestContext =>
-          val responder = createResponder(requestContext)
-          if(deleteAlbum(id))
-            responder ! UserDeleted
-          else
-            responder ! NodeNotFound("User")
+        delete { 
+          parameter('requesterId, 'auth) { (requesterId, auth) =>
+            requestContext =>
+            val responder = createResponder(requestContext)
+            if (authenticate(requesterId, auth)) {
+              if(deleteAlbum(id))
+                responder ! UserDeleted
+              else
+                responder ! NodeNotFound("User")
+            } else {
+              responder ! AuthFailed
+            }
+          }
         } ~
-        get { requestContext =>
-          var resultAlbum: Option[AlbumNode] = RestApi.albumList.find(_.id == id)
-          val responder = createResponder(requestContext)
-          resultAlbum.map(responder ! _.albumResponse())
-            .getOrElse(responder ! NodeNotFound("Album"))
+        get{
+          parameter('requesterId, 'auth) { (requesterId, auth) =>
+            requestContext =>
+            val responder = createResponder(requestContext)
+            if (authenticate(requesterId, auth)) {
+              var resultAlbum: Option[AlbumNode] = RestApi.albumList.find(_.id == id)
+              resultAlbum.map(responder ! _.albumResponse(requesterId))
+                .getOrElse(responder ! NodeNotFound("Album"))
+            } else {
+              responder ! AuthFailed
+            }
+          }
         }
       }
     } ~
     pathPrefix("friendsList") {
       pathEnd {
         post {
-          entity(as[FriendsList]) { friendsList => requestContext =>
+          entity(as[FriendsList]) { friendRequest => requestContext =>
             val responder = createResponder(requestContext)
-            val resultOwner: Option[UserNode] = RestApi.userList.find(_.id == friendsList.owner)
-            val resultFriend: Option[UserNode] = RestApi.userList.find(_.id == friendsList.friend)
+            val resultOwner: Option[UserNode] = RestApi.userList.find(_.id == friendRequest.owner)
+            val resultFriend: Option[UserNode] = RestApi.userList.find(_.id == friendRequest.friend)
             if(resultOwner.isEmpty) {
               responder ! NodeNotFound("User")
             } else if(resultFriend.isEmpty) {
               responder ! NodeNotFound("Friend")
             } else {
-              //both owner and friend exist, add in each others friends list.
-              resultOwner.get.friendsList = resultOwner.get.friendsList :+ friendsList.friend
-              resultFriend.get.friendsList = resultFriend.get.friendsList :+ friendsList.owner
-              println("Updated friends list for " + friendsList.owner + " and " + friendsList.friend)
-              responder ! FriendsListUpdated
+              if (authenticate(friendRequest.owner, friendRequest.auth)) {
+                //both owner and friend exist, check if they're not friends already
+                val alreadyFriend: Option[String] = resultOwner.get.friendsList.find(_ == friendRequest.friend)
+                if(alreadyFriend.isEmpty) {
+                  //not a friend yet
+                  //add in each others friends list.
+                  resultOwner.get.friendsList = resultOwner.get.friendsList :+ friendRequest.friend
+                  resultFriend.get.friendsList = resultFriend.get.friendsList :+ friendRequest.owner
+                  println("[SERVER] Updated friends list for " + friendRequest.owner + " and " + friendRequest.friend)
+                  responder ! FriendsListUpdated
+                } else {
+                  //already a friend
+                  responder ! FriendExists
+                }
+              } else {
+                responder ! AuthFailed
+              }
             }
           }
         }
@@ -281,6 +357,21 @@ trait RestApi extends HttpService with ActorLogging { actor: Actor =>
           responder ! userIdList
         }
       }
+    }~
+    pathPrefix("signNumber") {
+      path(Segment) { id =>
+        get { requestContext =>
+          val responder = createResponder(requestContext)
+          var resultUser: Option[UserNode] = RestApi.userList.find(_.id == id) //get user to store random number in user object
+          if(resultUser.isEmpty) {
+            responder ! NodeNotFound("User")
+          } else {
+            val rand : String = (new Random(System.currentTimeMillis)).nextInt(1000) + ""
+            resultUser.get.signNumber = rand
+            responder ! SignNumberRequest(rand)
+          }
+        }
+      }
     }
 
   private def createResponder(requestContext:RequestContext) = {
@@ -299,7 +390,6 @@ trait RestApi extends HttpService with ActorLogging { actor: Actor =>
   }
 
   private def deletePost(id:String): Boolean = {
-    println("delete user " + id)
     var resultPost: Option[PostNode] = RestApi.postList.find(_.id == id)
     if(resultPost.isEmpty) {
       return false
@@ -355,6 +445,36 @@ trait RestApi extends HttpService with ActorLogging { actor: Actor =>
     var resultUser: Option[UserNode] = RestApi.userList.find(_.id == friendListOfId)
     resultUser.get.friendsList = resultUser.get.friendsList.filterNot(_ == deleteUserId)
   }
+
+  def decryptAsym(text: String, key: PublicKey): String = {
+    val encryptedBytes: Array[Byte] = Base64.getDecoder().decode(text);
+    val cipher = Cipher.getInstance("RSA")
+    cipher.init(Cipher.DECRYPT_MODE, key)
+    try{
+      var textBytes:Array[Byte] = cipher.doFinal(encryptedBytes)
+      return new String(textBytes)
+    } catch {
+      case e: GeneralSecurityException => return "DecryptionFailed"
+    } 
+    
+  }
+
+  def stringToPublicKey(publicKey: String): PublicKey = {
+    var publicBytes: Array[Byte] = Base64.getDecoder().decode(publicKey);
+    val keySpec: X509EncodedKeySpec = new X509EncodedKeySpec(publicBytes);
+    val keyFactory: KeyFactory = KeyFactory.getInstance("RSA");
+    val key: PublicKey = keyFactory.generatePublic(keySpec);
+    return key
+  }
+
+  def authenticate(userId:String, text:String) : Boolean = {
+    val user = RestApi.userList.find(_.id == userId)
+    val publicKey:PublicKey = stringToPublicKey(user.get.publicKey)
+    val decryptedSign: String = decryptAsym(text, publicKey)
+    val authenticated: Boolean = decryptedSign.equals(user.get.signNumber)
+    user.get.signNumber = if (authenticated) "-1" else user.get.signNumber
+    authenticated
+  }
 }
 
 class Responder(requestContext:RequestContext) extends Actor with ActorLogging {
@@ -371,18 +491,29 @@ class Responder(requestContext:RequestContext) extends Actor with ActorLogging {
       killYourself
 
     case UserAlreadyExists =>
-      requestContext.complete(StatusCodes.Conflict, "A user with that email is already registered")
+    println("[SERVER] Email already exists.")
+      requestContext.complete(StatusCodes.Conflict, "[ERROR] A user with that email is already registered")
+      killYourself
+
+    case FriendExists =>
+    println("[SERVER] Friend already exists.")
+      requestContext.complete(StatusCodes.OK, "[ERROR] Requested user is already a friend.")
       killYourself
 
     case NodeNotFound(nodeType: String) =>
-      requestContext.complete(StatusCodes.NotFound, nodeType + " not found")  
-      killYourself    
+      println("[SERVER] Node not found " + nodeType)
+      requestContext.complete(StatusCodes.OK, "[ERROR] " + nodeType + " not found")  
+      killYourself
+
+    case SignNumberRequest(number: String) =>
+      requestContext.complete(StatusCodes.OK, number)
+      killYourself
 
     case response: PageResponse =>
       requestContext.complete(StatusCodes.OK, response)
       killYourself
 
-    case response: PostResponse =>
+    case Left(response: PostResponse) =>
       requestContext.complete(StatusCodes.OK, response)
       killYourself
 
@@ -401,6 +532,17 @@ class Responder(requestContext:RequestContext) extends Actor with ActorLogging {
     case response: Vector[String] =>
       requestContext.complete(StatusCodes.OK, response)
       killYourself
+
+    case Right(notAuthorizedError:String) =>
+      println("[SERVER] Not authorized to access requested node.")
+      requestContext.complete(StatusCodes.Unauthorized, notAuthorizedError)
+      killYourself
+
+    case AuthFailed => 
+      println("[SERVER] auth failed")
+      requestContext.complete(StatusCodes.OK, "[ERROR] Authentication Failed")
+      killYourself
+
    }
 
   private def killYourself = self ! PoisonPill
